tls: make tls_get_connection_info as ssize_t
'int' looks like too much as usual 0/-1 choice.

+++ usual/tls/tls.h
@@ -131,7 +131,7 @@ int tls_write(struct tls *_ctx, const void *_buf, size_t _buflen,
     size_t *_outlen);
 int tls_close(struct tls *_ctx);
 
-int tls_get_connection_info(struct tls *ctx, char *buf, size_t buflen);
+ssize_t tls_get_connection_info(struct tls *ctx, char *buf, size_t buflen);
 
 uint8_t *tls_load_file(const char *_file, size_t *_len, char *_password);
